requirements.txt
PyTorch
HuggingFace Transformers
Pillow
Numpy
